requirements.txt
amqp==2.5.0
billiard==3.6.0.0
boto3==1.9.233
botocore==1.12.237
braintree==3.56.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.0.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-guardian==2.0.0
django-paypal==1.0.0
django-rest-swagger==2.2.0
django-storages==1.7.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.15.2
easypost==4.0.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
paypalrestsdk==1.13.1
pika==0.12.0
Pillow==6.1.0
protobuf==3.10.0
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.4.2
python-freshdesk==1.2.7
pytz==2019.1
requests==2.22.0
rest-framework-generic-relations==1.2.2
rsa==4.0
s3transfer==0.2.1
sentry-sdk==0.12.3
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.35.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
gevent==20.9.0
redis==3.0.1
django-redis==4.12.1
redis-py-cluster
XlsxWriter==1.3.7
fontTools==4.22.0